More efficient checking of minimum severity level
We now implement `clog_log_channel` as a macro, which lets us test the
minimum severity level in code that generates the log messages.  This
lets that code perform the level check *before* setting up the stack
frame for the call to the log handlers.  With this change, we now only
need 3-4 machine instructions to determine if  a log message passes the
minimum severity level threshold.  This, in turn, lets us get rid of
`CLOG_TRACE_ENABLED`, and always compile trace messages into the
application.