Log message annotations
This patch implements the `clog_annotate_message` function, which is
used to add annotations to a log message.  The documentation states that
annotations are only visible to handlers “lower” in the stack (i.e.,
pushed onto the stack more recently) than the handler that creates the
annotation.  Our test case program verifies this behavior by pushing an
annotator onto the stack both before and after the handler that collects
the messages for verification.