Multiple annotations conversion
This patch adds a new format string conversion that prints all of the
annotations in a log message.  If there are no annotations, no text is
added to the result.  You provide an annotation spec that's used to
render each annotation.  The annotations are added in the same order as
they're generated by the log handlers.