CLOG environment variable
This patch adds a `clog_setup_logging` function, which programs can use
to create and register a logging handler with useful defaults.  The
default handler can be configured via environment variables.  Right now,
we only support one.  `CLOG` can be set to the name of a logging level,
and that will be used as the minimum logging level for the process.