CLOG_FORMAT environment variable
When using the default log handler, the user can set the `CLOG_FORMAT`
environment variable to override the format string used to render each
log message.  We still have the `clog_set_default_format` function so
that programs can provide their own default formats, which will be used
if the `CLOG_FORMAT` variable is not set.  (This can be useful if a
program uses annotation tags that should be included in the default log
format, for instance.)