Add structured events
You can now log structured events, which consists of a log message along
with a number of named fields.  We already had a nice framework that
allowed handlers to add annotations to log messages, which we can reuse
for these fields.  So in this implementation, a field is just an
annotation provided by the code generating the log message, instead of
by the handlers that receive the message.