Greatly reduce function call overhead
This patch makes the entire framework _much_ faster, by eliminating a
large number of function calls that were previously needed to handle a
message.  There's now a single callback method that handlers provide,
which processes the entirety of a message (including all annotations) in
one go.  And the macros that produce messages construct everything on
the stack, using inlined constructor functions, before handing the
message off for processing.

Handlers are still allowed to add additional fields to a message, and
it's still true that only "downstream" handlers are able to see the new
fields.  Before, we guaranteed that by explicitly skipping over the
"upstream" handlers in the function that was used to add new
annotations.  Now, it's guaranteed because of the invariant that the
message fields are instantiated on the stack — the handler create the
new fields in its `handle` callback, and is obligated to pop those
fields back off before returning.

All in all, we're at the point where we can process log messages
(including fields) with a null handler in ~10s of nanoseconds.