Use singly-linked list for message fields
We were only using a doubly-linked list before so that we could easily
iterate through a log message's fields in the same order that they
appear in the source code.  But we only have one place that really needs
to do that: the multi-field #*{} format specifier.  Everywhere else,
we're either saving results into an unordered container, or checking
each field against a single predicate.  For both of those, we can
iterate through the fields in any arbitrary order.

That means that we can simplify our field representation by using a
singly-linked list.  That cuts down on the size of each field in memory,
and on the number of instructions we need to initialize each one while
constructing a log message.

That means that we're now using an O(n^2) algorithm when we _do_ want to
iterate through the fields in order, but we won't really ever have a
large enough number of fields for that to make a difference.