Add macros for checking stash contents
This lets you use the same syntax to construct the "expected" message
when verifying the contents of a stash, that you use when logging the
messages in the first place.  This can be especially helpful if you have
any specialized field types and want to reuse them in your test cases.